Reorganized code files drafted some platform-specific constructs.
I decided to just make a library that'll be filled with the platform-
specific code I need to implement in order to get rendering working
on both Windows and Linux (and maybe others, in the future). I also
started implementing some of the function loading code for Windows
(already handled in Linux, thanks to my window manager project) as
well as a rendering window class (again, already handled in Linux,
but not yet included here).

OpenGL function and extension loading appears to be quite a bit
different in Windows compared to Linux. Evidently, one needs to
generate a dummy rendering context before loading the function
pointers, and then the real rendering context can be created.
However, even creating a dummy context requires a valid rendering
window to already exist, so there's a lot of setup that needs to
happen here (as opposed to Linux, where you can basically just
start loading stuff right out of the gate).

Unfortunately, because these two processes are reversed with respect
to one another in some important ways, creating a reusable and cross-
platform code base for both is going to be tricky.