Forced to add namespace wrappers around utility objects.
I'm pretty much being forced to wrap everything in this project in
a namespace to avoid collisions with the Windows API. This is
considered good practice in general, but namespace implementation
in C++ is clumsy at the very best, and since I don't (as a rule)
use a lot of outside dependencies, it was never an issue on Linux.

Now I have to go back and retrofit all of the other libraries for
combatibility with this change. Not a huge task, but a trivial one
that introduces additional boilerplate code requirements. I seriously
wish there were another language as powerful as C++ but without all
of the language design atrocities...