Added a bunch of documentation and removed constant pointers.
The constant pointers to the render target, shader pipeline, and
render window were going to prevent me from performing some
important tasks. So, for now, I'm removing them from the class
fields and setter methods. They will, however, remain for the
getter methods for the time being. This will prevent being able to
temporarily swap out any one of these and then replace it with the
old value without constant-casting, but for now this will have to
do until I can think of a better way doing that.