Removed the internal data store from this class.
I decided to convert this class into even more of an abstract base
or interface. This allows me to create subclasses with various types
of templated data storage without having to also keep all of the
OpenGL stuff in the header. It also lets me reuse most of the buffer
handling logic, which is mostly the same regardless of the underlying
storage class.