Reworked rendering context setup and added some documentation.
A day or two ago, I discovered that the way in which I was setting
up rendering contexts is incorrect on the Windows platform. In that
environment, OpenGL function pointers are directly related to the
rendering context in which they are loaded, meaning that any time
a new context gets created (i.e. when constructing new Window3D
instances), any function pointers loaded previously might become
invalid. This is a huge pain, so for now I'm going to operate on
the assumption that only one rendering context gets created per
running program instance. This shouldn't pose much of a problem,
though, given the behaviors I've seen in other rendering engines.