Fixed some shader code preprocessing bugs.
Firstly, I fixed a bug that tried to expand include macros even if
they were on commented lines. I also (partially, maybe) resolved a
subtle bug where null characters appear multiple times within the
string output of the 'Preprocess' method. My working theory is that
this is the result of concatenating multiple strings together through
the 'stringstream', which doesn't prevent the inputting of '\0'.
Thus, when I need to convert this from a non-null-terminated 'string'
to a 'const char*' when calling 'glShaderSource', the source code is
getting cutoff prematurely when it encounters an early null.