Major pipeline class updates to remove unnecessary functionality.
I went through and reorganized/purged a bunch of code in these
classes. Most of what was removed is functionality that is now being
handled by the 'GPU' class. I also moved shader program linking to
the 'GraphicsPipeline' class so that any subclasses can take advantage
of some common methods.