Heavily restructured this class.
I removed all inheritance from this class for now, although it's
looking like a good 'IBindable' candidate already. Instead of being
a specialized 'UniformMap', it now holds a per-entity 'UniformBuffer'
as well as a 'CommandBuffer', which was just implemented. Indexing
these buffers will be done through a 'std::map' that stores indices.