Finalized rotational properties and added documentation.
I ended up changing the name of the rotational data vector from
'_rotation' to '_orientation', mostly because I was out of valid
variable names and couldn't make a 'Rotation' property because of
the existing static constructor with that name. The new name still
preserves the variable's meaning and purpose, although admittedly
it is less ideal.