Improved 'UpdateSize' and renamed 'RenderArea' to 'ClientArea'.
I added in a check to 'UpdateSize' that prevents making changes to
cached data and triggering the 'OnResize' event when the size of
the window has not actually changed. This will help prevent any
unnecessary resource reallocations (e.g. the framebuffer) if this
method accidentally gets called.

I also renamed the internal area property to use the more common
'Client' descriptor. This should reflect the meaning of the property
a little better.

Lastly, I added some documentation to the new events and methods
of this class.