Had to remove use of the function 'memcpy'.
This started causing a weird exception to be thrown at one very
specific location (in a new class called 'VertexArray'). Substituting
manual array copying resolves the error, but honestly I don't
understand what the problem was in the first place and a quick Google
search on the matter didn't turn up anything useful.