Split the OpenGL implementation of 'VertexBuffer' into 'VertexArray'.
I created a new class called 'VertexArray' to solve a problem I've
been contemplating for a while now. This class contains all of the
OpenGL-related implementation code for creating vertex buffers on
the GPU. In so doing, I was then able to make 'VertexBuffer' a
templated class for the purpose of being able to swap in various
vertex structure definitions, without having to dump a bunch of
core graphics dependencies in the header file.

Unfortunately, at least to my eyes, this makes the buffer class
hierarchy a little janky, but I haven't been able to come up with a
better way of doing things while also keeping the major implementation
details separated from the interface. I'll have to think about this
one some more in the future.