Implemented prototyped support for indirect rendering of indexed geometry.
I ended up restructuring 'Scene3D' pretty heavily to accomplish
this, and that class in particular still has a long way to go. I
also decided to move the rendering logic over to there from the
'GPU' class. Part of me wanted all of that in 'GPU', but the more
I think about, the more it makes sense for that to live in 'Scene3D'.
Subclasses could then easily customize the rendering logic, and I
also planed on doing something analogous to the 'Scene3D' class for
compute pipelines, which would also benefit from customizable
execution methods.