Implemented application-side buffer element subset updating.
This one has been planned for a while now, but it was the
'BasicSphere' demo program that finally forced me to implement
it because performance was lousy when tessellating icosahedron
geometry more than twice. The bad performance was the result of
having to constantly recycle all of the command, uniform, and
vertex buffers any time there was any kind of a change to renderable
entities. Naturally, it wasn't a viable long-term solution.

This update only supports element subset updating of the 'PerEntity'
uniform buffers (i.e. entity coloring and spatial transformations).
In the future, I'll implement additional support for updating vertices,
indices, and commands (probably as the need arises). For the moment,
these updates still clear out and overwrite the entire buffer on the
GPU, which is potentially a performance problem, but that should be
relatively easy to fix. I just need to map and overwrite a range of
buffer data on the GPU.