Added a get-accessor for the rendering pipeline ID.
The idea here is to eventually get rid of the get-accessors that
are currently returning pointers (which could eventually lead to
all kinds of trouble). However, lots of things need information
like this from the GPU in order to function, so I need to think of
a sensible way to replace the free access that's currently there.