Forced implement a quick workaround for 'PathCommands' enumerators.
These values are of unsigned byte types and are meant to be placed
into arrays read by OpenGL and the GPU, meaning my usual approach
of just converting everything to integers fails here. This probably
means I need to make 'Enumerator' a templated class, but the number
of those that are already present in this library is becoming
alarmingly high. I need to do some thinking about what needs to be
done with all of this.