Began implementing SVG stroke rendering.
This isn't currently working properly, but the stroke is showing up
at least on the rounded-rectangle demo. I also removed all references
to string paths and the ability to specify paths this way. It's really
easy to do if I find I ever need it again, but path specification needs
to be done programmatically to be flexible enough for me.