Restructured setter properties, made most public methods virtual, and updated documentation.
I went through and restructured the property setters so that they
all filter down into common base set-accessors. Combined with the
addition of the 'virtual' keyword, this makes 'Transform' a better
base class for transformations that don't want to override every
single property that exists in order to modify behavior. Instead,
at least for setters, they now only need to override the base that
actually sets the data as opposed to the ones that just hand it off
to some other set-accessor.