Rewrote 'KeyboardKeys' to utilize a bitset instead of typical numeric bit flags.
I was forced to switch over to using a 'bitset' in order to solve
the lack of necessary storage capacity in integral types that normally
back enumerators. Although bitsets are slower, using them in this way
lets me keep the keyboard keys unified under one 'enum' structure
and, most importantly, lets me track the entire keyboard state in one
single variable. The enumerators themselves now function as indices
into the bitset instead of being the actual flags on which operations
are performed. So far, this approach appears to be working pretty
well.