Implemented better 3D first-person camera navigation.
Now, movements in the vertical direction are controlled via the
'Shift' and 'Space' keys, and the original view for the scene can
be restored via the 'F1' key.

I was going to quickly implement support for rendering a scene as
a wireframe only (hence the 'F2' key presence), but it turns out
that this is more difficult in OpenGL than it was in Direct3D,
where it was just a simple toggle setting somewhere. In OpenGL,
you're basically forced to implement that kind of rendering from
the ground up on your own. I definitely don't have time to do that
before the innovation sprint demonstration today.