Reset the tessellation count to '4'.
This demo will be used to showcase the power of indirect rendering
and its ability to mix with SVG rendering pretty seamlessly. Using
a tessellation count of '4' means that more than 1,200,000 vertices
are being processed and displayed on the screen every frame. Even
on my workstation GPU, framerates are totally stable above 60 Hz,
which I think is pretty impressive.