Refactored initialization code and successfully enabled native MSAA.
Finally, I've gotten MSAA to work, but only on the native framebuffer
supplied by the operating system (still have no idea what's wrong
with the user-defined framebuffer objects). That was much more
difficult than I anticipated, and basically required me to rewrite
almost all of the WGL-related code. Now I need to get anti-aliasing
working for ALL framebuffers.

Most of the changes here are intended to reorganize the in-source
test executable to look more like the renderers that back all of the
demo projects. The constructor here was starting to look a little
unruly.