Restructured window initialization to support MSAA context creation.
For now, I had to remove the dedicated window creation method and
move all of its code to the renderer's constructor to ensure that
things get initialized in the correct order. This is necessary to
support MSAA on native window-supplied framebuffers, which was very
difficult to get working properly. I'm working towards restoring
the original flexibility now, but for the time being this new code
layout is the only way to make it all work.