Implemented some dynamic path features.
These include a 'Clear' method for path commands, which clears out
any stored commands and associated coordinates, as well as a check
within the 'Add' method that converts any initial command into a
'Move' command. This was done to avoid drawing lines or curves that
begin at the origin instead of the first point that gets inserted.