Began creating a more unified texture class.
Most of the texturing commands are 1D/2D/3D agnostic, which means
having subclasses for each of them (times two in some cases for
multisampled textures) will result in a ton of very empty classes.
By using a handful of ugly switch/case blocks, I can remove the
need for almost all of them right away.

I'm keeping 'Texture2D' around for now because of its ability to
read in JPEG files, which is something that I may or may not
eventually include in 'Texture' as well.