Began implementing a general OpenGL rendering settings structure.
This will eventually allow for easier and more consistent state
changing via the GPU class. I also plan on incorporating this
structure into 3D scenes, although now that I'm thinking about it,
it may be necessary to define more interfaces and move some of this
code over to the OpenGL library (to prevent interdependencies from
forming). This is all still a work in progress for the moment.