Got a prototype implementation of staged SVG rendering working.
After a few days of mucking around with the interfaces, I finally
got SVGs to render using the new 'RenderStage2D' and 'SceneSVG'
classes. Once I've refined how this all works a little bit more,
this should break up most (if not all) of the major undesirable
inter-library dependencies. Afterwards, although pretty much
everything will (always) depend on the OpenGL and Utilities library,
the SVG and Platform libraries will be independent of one another,
and I'll have a clearer idea of how to produce new libraries for
the project in the future.