Refactored a large chunk of 'Entity3D' into the base class 'GeometricEntity3D'.
I'm working towards a full separation of the logic behind entity
appearances, behaviors, and rendering geometries. In the end, I'm
hoping that 3D renderable entities will become a composite of
characteristics that are essentially hot-swappable with one another
and can be stored completely separately from the entity itself.
Having a system like this would dramatically increase the flexibility
and modularity of the rendering engine as a whole.