Continuing work on rendering scene, stage draw buffer reorganization.
This is proving a lot trickier than I thought it would be. It's
starting to look like I need to implement a complete separation of
most entity-related resources sooner than I expected. By that, I
mean that essentially everything needs to become some abstracted or
opaque shared resource (e.g. geometry, styling, etc.).

That in turn would mean that I need a system for externally
controlling these resources sooner than expected. I had always
intended for that to be take the form of CSS files or to allow use
of some markup language. Either way, this would be a very complex
undertaking, and I'm really beginning to think that it's a necessary
next step, because I'm having a very hard time coming up with a
system for this that's compatible with the current entity setup.