Fixed a critical issue with 'EntityData' where additional padding was required and implemented equality checking operators.
Rendered scenes contained obvious errors for a little while due to
the fact that the new 'EntityData' structure wasn't propertly aligned
to 16 byte boundaries. To fix this, I added a 'Vector2' padding
element to the end of the structure and all is well again.

I also implemented equality checking for these data structures in
preparation for making them a shared resource among renderable
entities. This task is still very much a work in progress...