Implemented storage of keyboard state and improved directional movements.
Before, the direction of camera movements via keyboard use was fixed
until all keys were released, at which point the movement direction
could be updated. This behavior prevented being able to move in the
correct relative direction if the camera was also being rotated as
it moved. Now, camera movement direction is updated whenever the
camera is rotated, which is more consistent with standard first-
person camera behaviors.