Removed the restriction that stored elements be unique and implemented an 'IsValid' property for registry keys.
The uniqueness restriction was starting to get in the way and was
potentially very costly if applied to large data structures. Also,
there are several situations where I might want to register identical
structures separately. So, that behavior is gone, and anything can
now be inserted into a 'Registry' collection.