Exposed the 'Register' method for entities to the public.
Things were getting kind of weird while 'Register' was a protected
method of the interface. I hate that 'ResourceMapping' references
are now publicly visible, and will have to do something about that
in the near future, but for now this solves a bunch of problems that
arose when other libraries attempted to utilize this interface.