Re-implemented use of registry buffers for entity transformation data.
I had implemented and then removed this behavior a short while ago,
but with the recent removal of equality checks in the 'Registry'
class (which were no doubt expensive for transformation data), I'm
reinstating it. Use of registries lets me insert and remove data
elements without having to constantly manually refresh the indices
into the storage that I'm caching.