Prototyping some basic VM functionality in TypeScript.
Since TypeScript is becoming important at work and is closely related
to CoffeeScript, I figured this was a good opportunity to feel out
the language. To that end, I rewrote the 'Classes' file in TypeScript
(adding some new prototype classes in the process) and also began
implementing some primitive types that will eventually see extensive
use.