Continuing conceptualization of the new language 'Q'.
Although, to be fair, I'm getting pretty off track here. As it
stands, the virtual machine I've built should be capable of working
with rendering scenes, so I may go back to integrating that instead.
Creating a new programming language is an enormous project that I
probably don't have time for at the moment.