Deleted several of the separate shape entities.
This thins out the library a little bit (it's growing kind of big)
and will make it a little easier to develop a data-driven rendering
engine. I also deleted the old 'Space3D' class, since it was never
fully developed and I started working on a more general 'Octree'
class within the 'Utilities' library.