Updated to utilize the C++14 standard and to target the latest Windows SDK version.
This is the start of porting over some of my legacy code to the new
language standard (C++14) and Visual Studio 2017. I had added a compiler
flag to do this fairly recently, but was still using VS2015, which lacks
support for several C++14 features that are integrated in the 2017
version and that I'd very much like to use.