Completely restructured the 'IRenderable' interface.
I have some new ideas for the organization of this hierarchy that I want
to try, which make things a bit more compartmentalized and should help
out with writing a new scene graph structure.

'IRenderable' classes now contain only a model and associated meterial
data. Subcomponents have been removed and will be reimplemented as part
of the new scene graph structure. Other properties have been reorganized
into the model or material subcomponents and must now be accessed via
those.

The 'Model3D' class has been repurposed to serve only as a pairing of
3D geometry and a spatial transformation. It is also no longer derived
from the 'Entity3D' class, but is instead a component of it, and it lost
access to any material properties, which have been moved to 'Entity3D'.

Other changes include compatibility updates for this and other recent
commits. Notably, several files have been updated for the 'Transform' to
'Transform3D' rename and to take advantage of the new 'ITransformable3D'
interface/mixin.